+++ R/utils.R
@@ -139,6 +139,18 @@ assert_pid <- function(x) {
                             " is not a process id (integer scalar)"))
 }
 
+assert_grace <- function(x) {
+  if (is.integer(x) && length(x) == 1 && !is.na(x) && x >= 0) return(x)
+  if (is.numeric(x) && length(x) == 1 && !is.na(x) && x >= 0) {
+    xi <- as.integer(x)
+    # if x is non-zero, then return non-zero
+    if (xi == 0 && x > 0) return(1)
+    return(as.integer(x))
+  }
+  stop(ps__invalid_argument(match.call()$x,
+                            " is not a non-negative integer"))
+}
+
 assert_time <- function(x) {
   if (inherits(x, "POSIXct")) return()
   stop(ps__invalid_argument(match.call()$x,
@@ -157,6 +169,16 @@ assert_ps_handle_list <- function(x) {
     match.call()$x,
     " must be a process handle (ps_handle) or a list of process handles"
   ))
+}
+
+assert_ps_handle_or_handle_list <- function(p) {
+  if (!is.list(p)) {
+    assert_ps_handle(p)
+    p <- list(p)
+  } else {
+    assert_ps_handle_list(p)
+  }
+  p
 }
 
 assert_flag <- function(x) {